+++ tests/test_layers.py
@@ -1209,5 +1209,21 @@ def test_spec_augment_call(device):
         max_time_mask_proportion=max_time_mask_proportion,
         num_time_mask=num_time_mask,
         num_freq_mask=num_freq_mask,
-    )
+    ).to(device)
     spec_augment(feats, lengths)
+
+
+@pytest.mark.parametrize("mode", ["reflect", "constant", "replicate"])
+def test_random_shift_call(device, mode):
+    torch.manual_seed(50)
+    N, T, A, B = 50, 300, 13, 11
+    in_ = torch.rand(N, T, A, B, device=device)
+    in_lens = torch.randint(1, T + 1, (N,), device=device)
+    rand_shift = layers.RandomShift(1.0, mode).to(device)
+    out, out_lens = rand_shift(in_, in_lens)
+    assert out.dim() == 4
+    assert (out_lens >= in_lens).all()
+    assert out.size(0) == N
+    assert out.size(1) >= out_lens.max()
+    assert out.size(2) == A
+    assert out.size(3) == B